Make useFetch more robust and add param option to config
Previously, the `data` field could potentially send the string literal `'null'` as the `data` object, rather than leaving the `data` out. This commit fixes that by conditionally applying the data field if it's provided via the config instead.

Also additionally added this for query params, so that a consumer doesn't have to do `{axiosConfig: { params: "" }}`., which was verbose for no reason.